move state/util to util/
